+++ functions/Get-PSUnique.ps1
@@ -1,12 +1,13 @@
 ﻿
 Function Get-PSUnique {
     [cmdletbinding()]
-    [alias("gpsu")]
-    [OutputType("object")]
+    [alias('gpsu')]
+    [OutputType('object')]
     Param(
         [Parameter(Position = 0, Mandatory, ValueFromPipeline)]
         [ValidateNotNullOrEmpty()]
-        [object]$InputObject
+        [object]$InputObject,
+        [string[]]$Property
     )
 
     Begin {
@@ -16,7 +17,33 @@ Function Get-PSUnique {
     } #begin
 
     Process {
-        foreach ($item in $InputObject) {
+        if ($Property) {
+            foreach ($item in $InputObject) {
+                $props = $item.PSObject.Properties.where{ $_.name -in $Property }
+                if (-not $props) { continue }
+                if (-not $UniqueList.Exists({ -not (Compare-Object $args[0].PSObject.properties.where{ $_.name -in $Property }.value $props.value) })) {
+                    $UniqueList.add($item)
+                }
+            }
+        }
+        else {
+            foreach ($item in $InputObject) {
+                Try {
+                    if ($UniqueList.Exists( { -Not ( Compare-Object -ReferenceObject $args[0].PSObject.properties.value -DifferenceObject $item.PSObject.Properties.value )})) {
+                        Write-Debug "[$((Get-Date).TimeOfDay) PROCESS] Skipping: $($item |Out-String)"
+                    }
+                    else {
+                        Write-Debug "[$((Get-Date).TimeOfDay) PROCESS] Adding as unique: $($item | Out-String)"
+                        $UniqueList.add($item)
+                    }
+                }
+                Catch {
+                    Write-Warning "The input object can't be compared based on the number of properties. Try again using the Property parameter."
+                }
+
+            }
+        }
+        <# foreach ($item in $InputObject) {
             if ($UniqueList.Exists( { -not(Compare-Object $args[0].PSObject.properties.value $item.PSObject.Properties.value) })) {
                 Write-Debug "[$((Get-Date).TimeOfDay) PROCESS] Skipping: $($item |Out-String)"
             }
@@ -24,7 +51,7 @@ Function Get-PSUnique {
                 Write-Debug "[$((Get-Date).TimeOfDay) PROCESS] Adding as unique: $($item | Out-String)"
                 $UniqueList.add($item)
             }
-        }
+        } #>
     } #process
 
     End {